Abstracted out the video stream
